+++ Carbon.Environment/Functions/Test-CEnvVariable.ps1
@@ -1,0 +1,102 @@
+
+function Test-CEnvVariable
+{
+    <#
+    .SYNOPSIS
+    Tests if an environment variable exists.
+
+    .DESCRIPTION
+    The `Test-CEnvVariable` function tests if an environment variable exists. Pass the name of the variable to
+    the `Name` parameter (or pipe in multiple names). If a variable with that name exists in the current process,
+    returns `$true`. Otherwise, returns `$false`.
+
+    Use the `Scope` parameter to check if user-level or computer-level environment variables exist.
+
+    To check if a specific user has an environment variable, pass that user's credentials to the `Credential` parameter.
+
+    On Windows, environment variable names are case-insenstive. On Linux and macOS, they are case-sensitive.
+
+    .LINK
+    Remove-CEnvVariable
+
+    .LINK
+    Set-CEnvVariable
+
+    .EXAMPLE
+    Test-CEnvVariable -Name 'PATH'
+
+    Demonstrates how to check that an environment variable exists. In this case, will return `$true` if the `PATH`
+    environment variable exists at any scope.
+
+    .EXAMPLE
+    Test-CEnvVariable -Name 'MY_VAR' -Scope User
+
+    Demonstrates how to check that an environment variable exists at a specific scope. In this case, will return `$true`
+    if the user has a `MY_VAR` environment variable.
+
+    .EXAMPLE
+    'PATH' | Test-CEnvVariable
+
+    Demonstrates that you can pipe environment variable names to `Test-CEnvVariable`.
+    #>
+    [CmdletBinding(DefaultParameterSetName='CurrentUser')]
+    param(
+        # The name of the environment variable to check. Case-insensitive on Windows. Cse-sensitive on Linux and MacOS.
+        [Parameter(Mandatory, ValueFromPipeline, ValueFromPipelineByPropertyName)]
+        [String] $Name,
+
+        # The specific scope to check. By default, checks the current process.
+        [Parameter(ParameterSetName='CurrentUser')]
+        [EnvironmentVariableTarget] $Scope,
+
+        # The credential of the user whose user-level environment variables to check.
+        [Parameter(Mandatory, ParameterSetName='ForUser')]
+        [pscredential] $Credential
+    )
+
+    begin
+    {
+        Set-StrictMode -Version 'Latest'
+        Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+        $userEnvVars = [Collections.Generic.List[String]]::New()
+
+        if (-not $PSBoundParameters.ContainsKey('Scope'))
+        {
+            $Scope = [EnvironmentVariableTarget]::Process
+        }
+    }
+
+    process
+    {
+        if ($Credential)
+        {
+            $userEnvVars.Add($Name)
+            return
+        }
+
+        return ($null -ne [Environment]::GetEnvironmentVariable($Name, $Scope))
+    }
+
+    end
+    {
+        if ($Credential -and $userEnvVars.Count)
+        {
+            $parameters = $PSBoundParameters
+            [void]$parameters.Remove('Credential')
+            [void]$parameters.Remove('Name')
+            Start-Job -ScriptBlock {
+                    Import-Module -Name (Join-Path -path $using:moduleDirPath -ChildPath 'Carbon.Environment.psm1' -Resolve)
+                    $VerbosePreference = $using:VerbosePreference
+                    $ErrorActionPreference = $using:ErrorActionPreference
+                    $DebugPreference = $using:DebugPreference
+                    $WhatIfPreference = $using:WhatIfPreference
+                    $InformationPreference = $using:InformationPreference
+                    $using:userEnvVars | Test-CEnvVariable @using:parameters -Scope User
+                } -Credential $Credential |
+                Receive-Job -Wait -AutoRemoveJob |
+                Write-Output
+            return
+        }
+    }
+}